feat: implement user authentication and password recovery features
- Add email configuration for sending recovery codes.
- Create AuthResponseDTO and ResetPasswordDTO for handling authentication responses and password resets.
- Implement EmailService to send recovery codes via email.
- Update AuthContext to manage user authentication state and session loading.
- Refactor LoginScreen to handle user login with remember me functionality.
- Create RequestForgotPasswordScreen for requesting password reset codes.
- Implement UserScreenChangePassword for changing user passwords.
- Remove unnecessary components and screens related to password reset.
- Update API service to include authentication headers and user-related endpoints.
- Modify global styles for improved UI consistency.
- Update user-related types to support login and existing user structures.

+++ backend/src/main/java/com/example/schneider/trash_to_treasure/service/ItemService.java
@@ -2,35 +2,36 @@ package com.example.schneider.trash_to_treasure.service;
 
 import java.time.Instant;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.example.schneider.trash_to_treasure.dto.ItemDTO;
 import com.example.schneider.trash_to_treasure.entity.Item;
-import com.example.schneider.trash_to_treasure.entity.User;
 import com.example.schneider.trash_to_treasure.mapper.ItemMapper;
 import com.example.schneider.trash_to_treasure.repository.ItemFavoriteRepository;
 import com.example.schneider.trash_to_treasure.repository.ItemRepository;
-import com.example.schneider.trash_to_treasure.repository.UserRepository;
+import com.example.schneider.trash_to_treasure.security.CustomUserDetails;
 
 @Service
 public class ItemService {
 
-    private final ItemFavoriteRepository itemFavoriteRepository;
     private final ItemRepository itemRepository;
+    private final ItemFavoriteService itemFavoriteService;
     private final ItemMapper itemMapper;
     private final ImageService imageService;
+    private final UserService userService;
 
-    private final UserRepository userRepository;
-
-    public ItemService(ItemRepository itemRepository, ItemMapper itemMapper, ItemFavoriteRepository itemFavoriteRepository, UserRepository userRepository, ImageService imageService){
+    public ItemService(ItemRepository itemRepository, ItemMapper itemMapper, ItemFavoriteService itemFavoriteService, UserService userService, ImageService imageService){
         this.itemRepository = itemRepository;
         this.itemMapper = itemMapper;
-        this.itemFavoriteRepository = itemFavoriteRepository;
-        this.userRepository = userRepository;
+        this.itemFavoriteService = itemFavoriteService;
+        this.userService = userService;
         this.imageService = imageService;
     }
 
@@ -38,22 +39,25 @@ public class ItemService {
         List<Item> itemList = itemRepository.findAll();
 
         return itemList.stream()
-            .map(this::retrieveItemFavorites)
+            .map(itemFavoriteService::retrieveItemFavorites)
             .collect(Collectors.toList());
     }
 
     public ItemDTO.Response save(ItemDTO.Create itemCreateDTO, MultipartFile file){
+        Authentication auth =  SecurityContextHolder.getContext().getAuthentication();
+        CustomUserDetails user = (CustomUserDetails) auth.getPrincipal();
+
         Item itemSa = itemMapper.toEntity(itemCreateDTO);
         itemSa.setPhotoUrl(imageService.uploadImage(file));
         itemSa.setPostedAt(Instant.now());
-        Optional<User> user = userRepository.findById(1l);
-        itemSa.setPostedBy(user.get());
+        itemSa.setPostedBy(userService.getByIdOrThrow(user.getId()));
+
         return itemMapper.toDTO(itemRepository.save(itemSa));
     }
 
     public ItemDTO.Response findById(Long id){
-        Item item = itemRepository.findById(id).orElseThrow(() -> new RuntimeException("Item not found"));
-        return retrieveItemFavorites(item);
+        Item item = itemRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Item not found"));
+        return itemFavoriteService.retrieveItemFavorites(item);
     }
     
     public List<ItemDTO.Response> findByCordinates(Double latitudeUser, Double longitudeUser, Double area){
@@ -75,54 +79,58 @@ public class ItemService {
     }
 
     public List<ItemDTO.Response> findPostedByUser() {
-        Optional<User> user = userRepository.findById(1l);
-        return itemRepository.findByPostedBy(user).stream()
+        Authentication auth =  SecurityContextHolder.getContext().getAuthentication();
+        CustomUserDetails user = (CustomUserDetails) auth.getPrincipal();
+
+        return itemRepository.findByPostedBy(userService.getByIdOrThrow(user.getId())).stream()
                                                   .map(itemMapper::toDTO)
                                                   .collect(Collectors.toList());
     }
 
     public List<ItemDTO.Response> findCollectedByUser() {
-        Optional<User> user = userRepository.findById(1l);
-        return itemRepository.findByCollectedBy(user).stream()
+        Authentication auth =  SecurityContextHolder.getContext().getAuthentication();
+        CustomUserDetails user = (CustomUserDetails) auth.getPrincipal();
+
+        return itemRepository.findByCollectedBy(userService.getByIdOrThrow(user.getId())).stream()
                                                      .map(itemMapper::toDTO)
                                                      .collect(Collectors.toList());
     }
     
     public ItemDTO.Response update(Long id, ItemDTO.Update itemNewData){
         Item existingItem = itemRepository.findById(id)
-                                          .orElseThrow(() -> new RuntimeException("Item not found"));
+                                          .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Item not found"));
 
         itemMapper.updateEntityFromDTO(itemNewData, existingItem);
         return itemMapper.toDTO(itemRepository.save(existingItem));
     }
 
     public ItemDTO.Response collectItem(Long id) {
+        Authentication auth =  SecurityContextHolder.getContext().getAuthentication();
+        CustomUserDetails user = (CustomUserDetails) auth.getPrincipal();
 
         Item existingItem = itemRepository.findById(id)
-                                          .orElseThrow(() -> new RuntimeException("Item not found"));
-
-        System.out.println("Esse collected: " + existingItem.getCollectedBy());
+                                          .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Item not found"));
 
         if(existingItem.getCollectedBy() != null){
-            throw new RuntimeException("Item already taken");
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Item already taken");
         }
 
-        Optional<User> user = userRepository.findById(1l);
-        existingItem.setCollectedBy(user.get());
+        existingItem.setCollectedBy(userService.getByIdOrThrow(user.getId()));
 
         return itemMapper.toDTO(itemRepository.save(existingItem));
     }
 
     public void deleteById(Long id){
+        Authentication auth =  SecurityContextHolder.getContext().getAuthentication();
+        CustomUserDetails user = (CustomUserDetails) auth.getPrincipal();
+        
+        Item item = itemRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Item not found"));
+
+        if(!item.getPostedBy().equals(userService.getByIdOrThrow(user.getId()))) 
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Invalid Credentials");
+
+        itemFavoriteService.deleteByItem(item, userService.getByIdOrThrow(user.getId()));
+
         itemRepository.deleteById(id);
-    }
-
-    //Check and set if the item is favorite to the user or not
-    public ItemDTO.Response retrieveItemFavorites(Item item){
-        Optional<User> user = userRepository.findById(1l);
-
-        ItemDTO.Response dto = itemMapper.toDTO(item);
-        dto.setIsFavorite(itemFavoriteRepository.findByItemIdAndUserId(item.getId(), user.get().getId()).isPresent());
-        return dto;
     }
 }
